pyproject.toml
[tool.poetry]
name = "Djang"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
psycopg2 = "^2.9.7"
pillow = "^10.0.0"
ipython = "^8.15.0"
django-dump-load-utf8 = "^0.0.4"
image = "^1.5.33"
pytils = "^0.4.1"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
django-apscheduler = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
